Undo moving of tests
